cart-service/requirements.txt
# cart-service — Django REST Framework + PostgreSQL
Django==4.2.11
djangorestframework==3.15.1
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
gunicorn==21.2.0
python-dotenv==1.0.1